+++ CHANGELOG.md
@@ -5,5 +5,14 @@ All notable changes to the "ps-support" extension will be documented in this fil
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
 ## [Unreleased]
-
 - Initial release
+
+## [0.0.10] - 2025-06-28
+- Display PackScript errors in the Problems view
+- Activate extension only when `.packscript-ext.json` exists
+
+## [0.0.9] - 2025-06-28
+- Display PackScript stderr as a VS Code notification
+
+## [0.0.8] - 2025-06-28
+- Compile PackScript to the directory defined in `.packscript-ext.json` on file save